backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Video Processing
ffmpeg-python==0.2.0

# HTTP Client (for proxy testing)
httpx==0.25.1

# Environment Variables
python-dotenv==1.0.0

# Development Tools (optional)
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
